+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### Changed
-
 ## [Unreleased]
+
+### Changed
 
 - `dns-operator-openstack` now initially act on ClusterAPI `cluster` object to work with every
   infrastructure Provider via the `unstructured` client.
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reduce requeue time from five minutes to one minute to react faster to nginx IC being installed.
 - Improve finalizer addition&deletion to prevent unnecessary api calls.
 - Add new parameter to make RBAC configurable for different infra providers.
+- Renamed the project as `dns-operator-route53`.
 
 ## [0.4.1] - 2022-03-04