+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-08-10
+
 ### Changed
 
 - `dns-operator-openstack` now initially act on ClusterAPI `cluster` object to work with every
@@ -87,7 +89,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create api and ingress entries in Route53.
 
-[Unreleased]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/giantswarm/dns-operator-route53/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/giantswarm/dns-operator-route53/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/giantswarm/dns-operator-openstack/compare/v0.2.1...v0.3.0